Moved form constants to separate directories
