+++ app/javascript/src/components/Authentication/Login/index.jsx
@@ -6,29 +6,30 @@ import { Input } from "neetoui/v2/formik";
 import PropTypes from "prop-types";
 
 import authenticationApi from "apis/authentication";
-import formInitialValues from "constants/formInitialValues";
-import formValidationSchemas from "constants/formValidationSchemas";
+import { setAuthHeaders } from "apis/axios";
+import { useAuthDispatch } from "contexts/auth";
+import { useUserDispatch } from "contexts/user";
 
-const Signup = ({ history }) => {
+import { LOGIN_INITIAL_VALUES, LOGIN_VALIDATION_SCHEMA } from "./constants";
+
+const Login = ({ history }) => {
   const [loading, setLoading] = useState(false);
   const [submitted, setSubmitted] = useState(false);
 
-  const onSubmit = async formData => {
-    const { email, firstName, lastName, password, passwordConfirmation } =
-      formData;
+  const authDispatch = useAuthDispatch();
+  const userDispatch = useUserDispatch();
+
+  const onSubmit = async ({ email, password }) => {
     try {
       setLoading(true);
-      await authenticationApi.signup({
-        user: {
-          email,
-          first_name: firstName,
-          last_name: lastName,
-          password,
-          password_confirmation: passwordConfirmation,
-        },
-      });
-      Toastr.success("Successfully signed up. Please login.");
-      history.push("/login");
+      const {
+        data: { auth_token, user, is_admin },
+      } = await authenticationApi.login({ user: { email, password } });
+      authDispatch({ type: "LOGIN", payload: { auth_token, email, is_admin } });
+      userDispatch({ type: "SET_USER", payload: { user } });
+      setAuthHeaders();
+      history.push("/");
+      Toastr.success("Logged in successfully.");
     } catch (error) {
       logger.error(error);
     } finally {
@@ -40,51 +41,32 @@ const Signup = ({ history }) => {
     <div className="flex flex-row items-center justify-center w-screen h-screen p-6 overflow-x-hidden overflow-y-auto bg-gray-100">
       <div className="flex flex-col items-center justify-center w-full h-full mx-auto sm:max-w-md">
         <h2 className="mb-5 text-3xl font-extrabold text-center text-gray-800">
-          Signup
+          Sign In
         </h2>
         <Formik
-          initialValues={formInitialValues.signupForm}
+          initialValues={LOGIN_INITIAL_VALUES}
           validateOnBlur={submitted}
           validateOnChange={submitted}
           onSubmit={onSubmit}
-          validationSchema={formValidationSchemas.signupForm}
+          validationSchema={LOGIN_VALIDATION_SCHEMA}
         >
           {({ handleSubmit }) => (
             <Form className="w-full p-8 space-y-6 bg-white border rounded-md shadow">
               <Input
                 name="email"
                 type="email"
-                label="Email"
                 placeholder="oliver@example.com"
                 required
-              />
-              <Input
-                name="firstName"
-                type="text"
-                label="First name"
-                placeholder="Sam"
-                required
-              />
-              <Input
-                name="lastName"
-                type="text"
-                placeholder="Smith"
-                label="Last name"
-                required
+                label="Email"
+                data-cy="login-email-text-field"
               />
               <Input
                 name="password"
                 type="password"
+                placeholder="******"
+                required
                 label="Password"
-                placeholder="******"
-                required
-              />
-              <Input
-                name="passwordConfirmation"
-                type="password"
-                label="Confirm password"
-                placeholder="******"
-                required
+                data-cy="login-password-text-field"
               />
               <Button
                 type="submit"
@@ -93,25 +75,39 @@ const Signup = ({ history }) => {
                   setSubmitted(true);
                   handleSubmit();
                 }}
-                className="h-8"
                 loading={loading}
-                label="Signup"
                 fullWidth
+                className="h-8"
+                label="Login"
+                data-cy="login-submit-button"
               />
             </Form>
           )}
         </Formik>
-        <div className="flex flex-row items-center justify-start mt-4 space-x-1">
-          <p className="font-normal text-gray-600">Already have an account?</p>
-          <Button label="Login" style="link" to="/login" />
+        <div className="flex flex-col items-center justify-center mt-4 space-y-2">
+          <div className="flex flex-row items-center justify-start space-x-1">
+            <p className="font-normal text-gray-600">{`Don't have an account?`}</p>
+            <Button
+              label="Signup"
+              style="link"
+              to="/signup"
+              data-cy="sign-up-link"
+            />
+          </div>
+          <Button
+            label="Forgot password?"
+            style="link"
+            to="/my/password/new"
+            data-cy="forgot-password-link"
+          />
         </div>
       </div>
     </div>
   );
 };
 
-Signup.propTypes = {
+Login.propTypes = {
   history: PropTypes.object,
 };
 
-export default Signup;
+export default Login;
