--- app/javascript/src/constants/formInitialValues.js
@@ -1,28 +1,0 @@
-export default {
-  loginForm: {
-    email: "",
-    password: "",
-  },
-  resetPasswordForm: {
-    email: "",
-  },
-  signupForm: {
-    email: "",
-    firstName: "",
-    lastName: "",
-    password: "",
-    passwordConfirmation: "",
-  },
-  notesForm: {
-    title: "",
-    description: "",
-    assignedContact: "",
-    tags: "",
-  },
-  contactsForm: {
-    firstName: "",
-    lastName: "",
-    email: "",
-    role: "",
-  },
-};
